changing close error helper to take long ints

+++ 0x15-file_io/3-cp.c
@@ -6,9 +6,9 @@
  * @fd: file descriptor that failed to close
  */
 
-void close_error(int fd)
+void close_error(long int fd)
 {
-	dprintf(STDERR_FILENO, "Can't close fd %d\n", fd);
+	dprintf(STDERR_FILENO, "Can't close fd %li\n", fd);
 	exit(100);
 }
